requirements.txt
streamlit
fasttext-wheel
requests